major sepobject update-> now a photomapgit checkout Also High level collection will now be stored in the 'collections' directory linkl instruments are
