+++ astrobject/__init__.py
@@ -1,16 +1,22 @@
 #
 # Init the astro objects
 #
-import baseobject
-import spectroscopy 
-import transient 
-import photometry   
-#import multiband
-import collection
-import survey
-import instruments
-import astrometry
 
+# - fundamentals
+import baseobject
+import photometry 
+import astrometry
+import collection
+import spectroscopy # not quite used yet
+# - directory of tools
+import instruments
+import collections
+
+import transient # No used yet
+import survey # No used yet
+
+
+from collections import *
 from instruments.instrument import *
 from photometry import *
 from baseobject import *
