+++ astrobject/collections/__init__.py
@@ -1,0 +1,10 @@
+#
+# Collection from the Astrobject
+#
+import photospatial
+from photospatial import *
+#import photospectral
+#import phototemporal
+
+
+
