+++ astrobject/photometry.py
@@ -97,55 +97,6 @@ def photopoint(lbda=None,flux=None,var=None,
                       mjd=mjd,zp=zp,bandname=bandname,
                       **kwargs)
 
-def lightcurve(datapoints,times,**kwargs):
-    """
-    This functions enables to create a lightcurve object
-    from the given datapoints and times. The datapoints can
-    either be a list of 'PhotoPoint' or a dictionnary having
-    {flux, variance, lbda_s}
-
-    Parameters
-    ----------
-    datapoints: [list/dict]        Two formes fo data can be given to this function.
-
-                                   1) datapoints = list-of-PhotoPoint
-                                   This is a simple list (or array) of astrobject
-                                   PhotoPoint. Its size must corresponds to 'times' one
-                                   
-                                   2) a dictionnary having 3 parameters: lbda, fluxes,
-                                   and variances. Each have to have the size of 'times'
-                                   except lbda that could be simple float since all
-                                   must share the same lbda.
-
-    times: [array]                 This is the list/array of times corresponding to
-                                   each data points.
-
-    - options -
-
-    Return
-    ------
-    LightCurve
-    
-    """
-    if type(datapoints) is dict:
-        # this must be a dictsource:
-        datapoints = dictsource_2_photopoints(datapoints)
-        
-    if type(datapoints) is list or type(datapoints) is np.ndarray:
-        # - This should be a list of PhotoPoints, otherwise create will break
-        if len(times) != len(datapoints):
-            raise ValueError("'times and datapoints must have the same size")
-        
-        return LightCurve(datapoints, times)
-    
-    raise TypeError("'datapoints' must be a list of photopoints or a dictsource")
-        
-
-def sexobjects(sexoutput):
-    """
-    """
-    return SexObjects(sexoutput)
-    
 # ========================== #
 #  Internal Tool             #
 # ========================== #
@@ -204,9 +155,6 @@ def dictsource_2_photopoints(dictsource,**kwargs):
             for flux,var,lbda,source,ins in zip(fluxes,variances,lbda,
                                                 source,instrument_name)]
 
-    
-    
-    
 #######################################
 #                                     #
 # Base Object Classes: Image          #
@@ -912,19 +860,23 @@ class Image( BaseObject ):
         Void [or ndarray(sep.extract output) is returnobjects set to True]
         """
         from sep import extract
+        from collections import get_sepobject
         if thresh is None:
             thresh = self._get_sep_extract_threshold_()
             
         o = extract(self.data,thresh,**kwargs)
-        self._derived_properties["sepobjects"] = sexobjects(o)
+        self._derived_properties["sepobjects"] = get_sepobject(o) 
 
         if self.has_wcs():
             self.sepobjects.set_wcs(self.wcs)
 
         if set_catalogue and self.has_catalogue():
-            self.sepobjects.set_catalogue(self.catalogue,force_it=True)
+            # by give the catalogue and not a copy, the matching information
+            # is consistant between the sepobject and this instance (same catalogue)
+            self.sepobjects.set_catalogue(self.catalogue, reset=False)
             if match_catalogue:
                 self.sepobjects.match_catalogue()
+                
                     
         if returnobjects:
             return self.sepobjects
@@ -1701,9 +1653,9 @@ class PhotoPoint( BaseObject ):
         # ****************** #
         # * Creation       * #
         # ****************** #
-        self._properties["lbda"] = np.float(lbda)
+        self._properties["lbda"] = np.float(lbda) if lbda is not None else None
         self._properties["flux"] = np.float(flux)
-        self._properties["var"]  = np.float(var)
+        self._properties["var"]  = np.float(var) if var is not None else np.NaN
 
         self._side_properties["source"] = source
         self._side_properties["instrument_name"] = instrument_name
@@ -1822,10 +1774,11 @@ class PhotoPoint( BaseObject ):
     
     @mjd.setter
     def mjd(self,value):
-        if value < 48987:
-            print "Information mjd prior to 1993..."
-        elif value > 58849.0:
-            print "Information mjd posterior to 2020..."
+        if value is not None:
+            if value < 48987:
+                print "Information mjd prior to 1993..."
+            elif value > 58849.0:
+                print "Information mjd posterior to 2020..."
         
         self._properties["mjd"] = value
 
@@ -1835,10 +1788,11 @@ class PhotoPoint( BaseObject ):
     
     @zp.setter
     def zp(self,value):
-        if value <0:
-            raise ValueError("a zp cannot be negative")
-        if value >35:
-            warnings.warn("(PhotoPoint instance) the given zp is above 35...")
+        if value is not None:
+            if value==value and value <0:
+                raise ValueError("a zp cannot be negative")
+            if value==value and value >35:
+                warnings.warn("(PhotoPoint instance) the given zp is above 35...")
         
         self._properties["zp"] = value
         
@@ -1848,8 +1802,10 @@ class PhotoPoint( BaseObject ):
     
     @bandname.setter
     def bandname(self,value):
-        if type(value) != str and type(value) != np.string_:
-            raise TypeError("The bandname must be a string", type(value))
+        if value is not None:
+            if type(value) != str and type(value) != np.string_:
+                raise TypeError("The bandname must be a string", type(value))
+        
         self._properties["bandname"] = value
 
     @property
@@ -2116,676 +2072,3 @@ class LightCurve( BaseObject ):
     def zps(self):
         # - a unique lbda is tested while loading so this should be always the same
         return [p.zp for p in self.photopoints]
-
-#######################################
-#                                     #
-# Base Object Classes: SEPObjects     #
-#                                     #
-#######################################
-class SexObjects( BaseObject ):
-    """This instance parse the ourput from Sextractor/SEP and have
-    convinient associated functions"""
-
-    _properties_keys = ["data"]
-    _side_properties_keys = ["catalogue","wcs"]
-    _derived_properties_keys = ["catmatch","starmask"]
-    
-    def __init__(self,sexoutput=None,wcs=None,empty=False):
-        """
-        = Load the SexObject =
-
-        Parameters
-        ----------
-
-        sexoutput: [ndarray/file]  The ndarray as output by sextractor/sep or the
-                                   associated file
-
-        - options -
-
-        empty: [bool]              return an empy object
-        """
-        self.__build__()
-        if empty:
-            return
-
-        self.create(sexoutput)
-        if wcs is not None:
-            self.set_wcs(wcs)
-
-            
-    # ====================== #
-    # = Main Methods       = #
-    # ====================== #
-    def create(self,sexoutput,force_it=False):
-        """
-        """
-        if self.has_data() and force_it is False:
-            raise AttributeError("'data' is already defined."+\
-                    " Set force_it to True if you really known what you are doing")
-                    
-        sexdata = self._read_sexoutput_input_(sexoutput)
-        if sexdata is None:
-            warning.warn("(SexObjects.create) empty input data. *Empty SexObjects loaded*")
-            return
-        
-        # ****************** #
-        # * Creation       * #
-        # ****************** #
-        self._properties["data"] = sexdata
-
-        
-    def match_catalogue(self,catalogue=None,force_it=False,arcsec_size=2):
-        """
-        This methods enable to attached a given sexobject entry
-        to a catalog value.
-        You can set a catalogue.
-        """
-        # --------------
-        # - input 
-        if catalogue is not None:
-            self.set_catalogue(catalogue,force_it=force_it)
-
-        if not self.has_catalogue():
-            raise AttributeError("No 'catalogue' defined or given.")
-        
-        if self.has_wcs():
-            # -- matching are made in degree space
-            idxcatalogue, idxsexobjects, d2d, d3d = self.sky_radec.search_around_sky(self.catalogue.sky_radec, arcsec_size*units.arcsec)
-        else:
-            raise NotImplementedError("You currently need a wcs solution in the SexObjects to match a catalogue")
-        # --------------------
-        # - Save the results
-        self._derived_properties["catmatch"] = {
-            "idx_catalogue":idxcatalogue,
-            "idx":idxsexobjects,
-            "angsep":d2d
-            }
-        
-        self.catalogue.set_matchedmask(idxcatalogue)
-        self._derived_properties["starmask"] = np.asarray([i in self.catmask[self.catstarmask]
-                                                for i in range(len(self.data["x"]))],dtype=bool)
-
-    def idx_to_mask(self,idx):
-        """
-        Change the list of index to a True/False mask: index in the list
-        are the True values
-
-        Return:
-        -------
-        mask (boolean array)
-        """
-        mask = np.zeros(self.nobjects,dtype=bool)
-        for i in idx:
-            mask[i] = True
-        return np.asarray(mask,dtype=bool)
-        
-    # ------------------ #
-    # - SETTER         - #
-    # ------------------ #
-    def set_wcs(self,wcs,force_it=False):
-        """
-        """
-        if self.has_wcs() and force_it is False:
-            raise AttributeError("'wcs' is already defined."+\
-                    " Set force_it to True if you really known what you are doing")
-
-        self._side_properties["wcs"] = astrometry.get_wcs(wcs)
-        
-
-    def set_catalogue(self,catalogue,force_it=True,
-                      default_isolation_def = 10*units.arcsec):
-        """
-        Parameters
-        ---------
-
-        default_isolation_def: [ang dist]
-                                   If 'define_around' has not be ran yet,
-                                   this scale will be used.
-                                   This is important to know which object is
-                                   isolated.                                   
-                                   
-        Return
-        ------
-        Voids
-        """
-        if self.has_catalogue() and force_it is False:
-            raise AttributeError("'catalogue' already defined"+\
-                    " Set force_it to True if you really known what you are doing")
-
-        if "__nature__" not in dir(catalogue) or catalogue.__nature__ != "Catalogue":
-            raise TypeError("the input 'catalogue' must be an astrobject catalogue")
-
-        # -------------------------
-        # - Add the world_2_pixel
-        if not catalogue.has_wcs():
-            warnings.warn("WARNING the given 'catalogue' has no pixel coordinates. Cannot load it")
-            return
-        
-        if catalogue.nobjects_in_fov < 1:
-            warnings.warn("WARNING No object in the field of view, catalogue not loaded")
-            return
-
-        if not catalogue._is_around_defined():
-            catalogue.define_around(default_isolation_def)
-        self._side_properties["catalogue"] = catalogue
-                
-    # ------------------ #
-    # - GETTER         - #
-    # ------------------ #
-    def get(self,key,mask=None):
-        """
-        This function enable to get from the data the values of the given keys
-        or derived values, like ellipticity. Set 'help' for help.
-
-        *Remark* 'key' could be an list of keys.
-
-        Returns:
-        --------
-        array (or list of)
-        """
-        if not self.has_data():
-            raise AttributeError("no 'data' defined")
-
-        if "__iter__" in dir(key):
-            return [self.get(key_,mask=mask) for key_ in key]
-        
-        # -- These are the default key values
-        _data_keys_ = self.data.keys()
-        _matching_keys_ = ["angsep"]
-        _derived_keys_ = ["elongation","ellipticity"]
-        help_text = " Known keys are: "+", ".join(_data_keys_+_matching_keys_+_derived_keys_)
-
-        if key in ["help","keys","keylist"]:
-            print help_text
-            return
-        
-        # -- These are from the data
-        if key in _data_keys_:
-            val_ = self.data[key]
-        # -- These are the catalogue values
-        elif key in _matching_keys_:
-            if not self.has_catmatch():
-                raise AttributeError("no 'catmatch' defined. The matching has not been ran")
-            val_ = self.catmatch[key]
-        # -- These are derived values
-        elif key in _derived_keys_:
-            if key == "elongation":
-                val_ = self.get("a") / self.get("b")
-            elif key == "ellipticity":
-                val_ = 1. - 1. / self.get("elongation")
-        else:
-            raise ValueError("Cannot parse '%s'."%key +\
-                          help_text)
-                          
-        return val_ if mask is None else val_[mask]
-
-    def get_fwhm_pxl(self,stars_only=True,isolated_only=True,
-                    catmag_range=[None,None]):
-        """
-        This is defined as 2 * sqrt(ln(2) * (a^2 + b^2))
-        """
-        if not self.has_catmatch():
-            warnings.warn("No Catalogue matching loaded. No masking available")
-            warnings.warn("With no catalogue masking, the derived fwhm might not be accurate")
-            print "WARNING - With no cataloguemasking, the derived fwhm might not be accurate"
-            mask = None
-        else:
-            mask = self.get_indexes(isolated_only=isolated_only,
-                                    stars_only=stars_only,
-                                    catmag_range=catmag_range)
-            
-        return np.median(2 * np.sqrt( np.log(2) * (self.get('a',mask)**2 + \
-                                                   self.get('b',mask)**2)))
-    
-    # -----------------
-    # - get ellipse    
-    def get_ellipse_mask(self,width,height, r=3, apply_catmask=False):
-        """
-        This method returns a boolean mask of the detected ellipses
-        on the given width x height pixels image
-
-        (this method is based on the sep mask_ellipse function)
-        
-        Parameters:
-        -----------
-        r: [float]                 The scale of the ellipse (r=1 is a typical
-                                   contours included the object ; 2 enables to
-                                   get the tail of most of the bright sources
-
-        apply_catmask: [bool]      Only mask the detected object associated with
-                                   the current catalogue. If no catalogue loaded,
-                                   this will be set to False in any case.
-
-        Returns:
-        -------
-        2D-bool array (height x width)
-        """
-        from sep import mask_ellipse
-        ellipsemask = np.asarray(np.zeros((height,width)),dtype="bool")
-        mask = None if not apply_catmask else self.catmask
-        # -- Apply the mask to falsemask
-        mask_ellipse(ellipsemask,
-                     self.get('x',mask=mask),self.get('y',mask=mask),
-                     self.get('a',mask=mask),self.get('b',mask=mask),
-                     self.get('theta',mask=mask),
-                     r=r)
-        
-        return ellipsemask
-
-    def get_detected_ellipses(self,scaleup=5,apply_catmask=True,
-                stars_only=False, isolated_only=False,
-                catmag_range=[None,None]):
-        """
-        Get the matplotlib Patches (Ellipse) defining the detected object. You can
-        select the returned ellipses using the apply_catmask, stars_only,
-        isolated_only and catmag_range cuts.
-
-        'scaleup' scale the radius used of the ellipses. 5 means that most of the
-        visible light will be within the inside the returned ellipses.
-
-        Return
-        ------
-        list of patches
-        """
-        if not self.has_data():
-            print "WARNING [Sexobjects] No data to display"
-            return
-        
-        from matplotlib.patches import Ellipse
-        
-        # -- maskout non matched one if requested
-        if not self.has_catalogue():
-            apply_catmask = False
-            
-        mask = None if not apply_catmask else\
-          self.get_indexes(isolated_only=isolated_only,stars_only=stars_only,
-                        catmag_range=catmag_range)
-            
-        # -------------
-        # - Properties
-        return [Ellipse([x,y],a*scaleup,b*scaleup,
-                        t*units.radian.in_units("degree"))
-                for x,y,a,b,t in zip(self.get("x",mask=mask),self.get("y",mask=mask),
-                                     self.get("a",mask=mask),self.get("b",mask=mask),
-                                     self.get("theta",mask=mask) )]
-    
-    def get_median_ellipse(self,mask=None,clipping=[3,3]):
-        
-        """This methods look for the stars and return the mean ellipse parameters"""
-        if not self.has_catalogue():
-            apply_catmask = False
-          
-        # -- apply the masking
-        a_clipped,_alow,_ahigh = sigmaclip(self.get("a",mask=mask),*clipping)
-        b_clipped,_blow,_bhigh = sigmaclip(self.get("b",mask=mask),*clipping)
-        t_clipped,_tlow,_thigh = sigmaclip(self.get("theta",mask=mask),*clipping)
-        # - so        
-        psf_a,psf_b,psf_t = a_clipped.mean(),b_clipped.mean(),t_clipped.mean()
-        m = np.sqrt(len(a_clipped)-1)
-        
-        return [psf_a,np.std(a_clipped)/m],[psf_b,np.std(t_clipped)/m],\
-        [psf_t,np.std(t_clipped)/m]
-
-        
-    # ---------------- #
-    # - get Mask     - #
-    # ---------------- #
-    def get_mask(self,isolated_only,stars_only,
-                 catmag_range=[None,None]):
-        """
-        This main masking method builds a boolean array following
-        the requested cuts. Remark that this implies doing a catalogue cuts
-        since stars and magnitude information arises from the catalogue.
-
-        Returns
-        -------
-        array (dtype=bool)
-        """
-        if catmag_range[0] is None:
-            warnings.warn("No lower catalogue magnitude limit given. 13mag set.")
-            catmag_range[0] = 13
-            
-        mask = np.asarray(self.get_catmag_mask(*catmag_range))
-        if isolated_only:
-            mask = mask & self.catisolatedmask
-        if stars_only:
-            mask = mask & self.catstarmask
-        
-        return mask
-
-    def get_catmag_mask(self,magmin,magmax):
-        """
-        return the boolen mask of which matched point
-        belong to the given magnitude range.
-        Set None for no limit
-
-        (see also get_mask)
-        
-        Returns
-        -------
-        array (dtype=bool)
-        """
-        if not self.has_catalogue():
-            raise AttributeError("no 'catalogue' loaded")
-        if not self.has_catmatch():
-            raise AttributeError("catalogue has not been matched to the data")
-    
-        mags = self.catalogue.mag[self.catmatch["idx_catalogue"]]
-        magmin = np.min(mags) if magmin is None else magmin
-        magmax = np.max(mags) if magmax is None else magmax
-        return (mags>=magmin) & (mags<=magmax)
-
-    def get_indexes(self,isolated_only=False,stars_only=False,
-                    catmag_range=[None,None], cat_indexes=False):
-        """
-        Converts mask into index. Particularly useful to access the catalogue
-        values (set cat_indexes to True)
-        
-        Returns
-        -------
-        array of indexes
-        """
-        id = "idx_catalogue" if cat_indexes else "idx"
-        return np.unique(self.catmatch[id][self.get_mask(isolated_only=isolated_only,
-                                                  stars_only=stars_only,
-                                                  catmag_range=catmag_range
-                                                  )])
-
-    def get_photomap(self, matched_only=True,
-                     stars_only=False, isolated_only=False):
-        """
-        """
-        print "to be done"
-
-    # ------------------- #
-    # - PLOT Methods    - #
-    # ------------------- #    
-    # - Display
-    def display(self,ax,draw=True,
-                apply_catmask=True,
-                stars_only=False, isolated_only=False,
-                catmag_range=[None,None]):
-        """
-        """
-        ells = self.get_detected_ellipses(scaleup=5,apply_catmask=apply_catmask,
-                                          stars_only=stars_only,
-                                          isolated_only=isolated_only,
-                                          catmag_range=catmag_range)
-
-        for ell in ells:
-            ell.set_clip_box(ax.bbox)
-            ell.set_facecolor("None")
-            ell.set_edgecolor("k")
-            ax.add_patch(ell)
-        if draw:
-            ax.figure.canvas.draw()
-
-
-    # - Histograms
-    def show_hist(self,toshow="a",ax=None,
-                  savefile=None,show=True,
-                  apply_catmask=True,
-                  stars_only=False, isolated_only=False,
-                  catmag_range=[None,None],
-                  **kwargs):
-        """ Show the histogram of any given key """
-        # -- Properties -- #
-        if not self.has_catalogue():
-            apply_catmask = False
-            
-        mask = None if not apply_catmask else\
-          self.get_indexes(isolated_only=isolated_only,stars_only=stars_only,
-                        catmag_range=catmag_range)
-        
-        v = self.get(toshow)[mask]
-
-        # -- Setting -- #
-        from ..utils.mpladdon import figout
-        import matplotlib.pyplot as mpl
-        self._plot = {}
-        
-        if ax is None:
-            fig = mpl.figure(figsize=[8,6])
-            ax  = fig.add_axes([0.1,0.1,0.8,0.8])
-        elif "hist" not in dir(ax):
-            raise TypeError("The given 'ax' most likely is not a matplotlib axes. "+\
-                             "No imshow available")
-        else:
-            fig = ax.figure
-
-        # -- Properties -- #
-        default_prop = dict(histtype="step",fill=True,
-                            fc=mpl.cm.Blues(0.7,0.5),ec="k",lw=2)
-        prop = kwargs_update(default_prop,**kwargs)
-
-        out = ax.hist(v,**prop)
-
-        self._plot['ax'] = ax
-        self._plot['fig'] = fig
-        self._plot['hist'] = out
-        self._plot['prop'] = kwargs
-        fig.figout(savefile=savefile,show=show)
-
-    # - Ellipse
-    def show_ellipses(self,ax=None,
-                      savefile=None,show=True,
-                      apply_catmask=True,stars_only=False,
-                      isolated_only=False,catmag_range=[None,None],
-                      **kwargs):
-        """
-        """
-        if not self.has_data():
-            print "WARNING [Sexobjects] No data to display"
-            return
-        
-        from matplotlib.patches import Ellipse,Polygon
-        from ..utils.mpladdon import figout
-        import matplotlib.pyplot as mpl
-        self._plot = {}
-        # ------------------- #
-        # - axes            - #
-        # ------------------- #
-        if ax is None:
-            fig = mpl.figure(figsize=[6,6])
-            ax  = fig.add_axes([0.1,0.1,0.8,0.8])
-        elif "hist" not in dir(ax):
-            raise TypeError("The given 'ax' most likely is not a matplotlib axes. "+\
-                             "No imshow available")
-        else:
-            fig = ax.figure
-
-        # ------------------- #
-        # - axes            - #
-        # ------------------- #
-        if not self.has_catalogue():
-            apply_catmask = False
-            
-        mask = None if not apply_catmask else\
-          self.get_indexes(isolated_only=isolated_only,stars_only=stars_only,
-                        catmag_range=catmag_range)
-        # -------------
-        # - Properties
-        ells = [Ellipse([0,0],2.,2*b/a,t*units.radian.in_units("degree"))
-                for a,b,t in zip(self.get("a",mask=mask),self.get("b",mask=mask),
-                                 self.get("theta",mask))]
-        # -- Show the typical angle
-        psf_a,psf_b,psf_theta = self.get_median_ellipse(mask=mask)
-        ellipticity = 1- psf_b[0]/psf_a[0]
-        # - cos/ sin what angle in radian
-        
-        ax.plot([0,np.cos(psf_theta[0])*ellipticity],
-                [0,np.sin(psf_theta[0])*ellipticity],ls="-",lw=2,
-                 color=mpl.cm.Blues(0.8),zorder=8)
-
-        Cone_error = Polygon( [ [0,0],[np.cos(psf_theta[0]-psf_theta[1])*ellipticity,
-                                       np.sin(psf_theta[0]-psf_theta[1])*ellipticity],
-                                [np.cos(psf_theta[0]+psf_theta[1])*ellipticity,
-                                 np.sin(psf_theta[0]+psf_theta[1])*ellipticity],
-                                 [0,0]]
-                                )
-        Cone_error.set_facecolor(mpl.cm.Blues(0.8))
-        Cone_error.set_edgecolor(mpl.cm.Blues(0.8))
-        Cone_error.set_alpha(0.3)
-        ax.add_patch(Cone_error)
-
-        # -- Show the Ellipses
-        for ell in ells:
-            ell.set_clip_box(ax.bbox)
-            ell.set_facecolor("None")
-            ell.set_edgecolor("k")
-            ell.set_alpha(0.1)
-            ax.add_patch(ell)
-
-        ax.set_ylim(-1.1,1.1)
-        ax.set_xlim(-1.1,1.1)
-        # -- show the center
-        ax.axvline(0,ls="--",color="k",alpha=0.2)
-        ax.axhline(0,ls="--",color="k",alpha=0.2)
-        
-        self._plot['ax'] = ax
-        self._plot['fig'] = fig
-        self._plot['prop'] = kwargs
-        fig.figout(savefile=savefile,show=show)
-        
-    # =========================== #
-    # Internal Methods            #
-    # =========================== #
-    def _read_sexoutput_input_(self,sexoutput):
-        """ Parse the input - for flexible inputs"""
-
-        # ---------------
-        # -- no input
-        if sexoutput is None:
-            return
-        
-        # ---------------
-        # -- data input
-        if type(sexoutput) == np.ndarray:
-            # let's check
-            try:
-                _ = sexoutput["cxx"]
-            except :
-                raise TypeError("the given 'sexoutput' ndarray has no 'cxx' key."+\
-                                "\n"+" It most likely is not a sextrator/sep output ")
-            return Table(sexoutput)
-        
-        # ---------------
-        # -- file input
-        if type(sexoutput) == str:
-            if sexoutput.endswith(".pkl"):
-                from ..utils.tools import load_pkl
-                self._read_sexoutput_input_(load_pkl(sexoutput))
-                
-            # -- So this is a sextrator output ?
-            return
-
-    # =========================== #
-    # Properties and Settings     #
-    # =========================== #
-    # ----------------      
-    # -- Data
-    @property
-    def data(self):
-        return self._properties["data"]
-    
-    def has_data(self):
-        return False if self.data is None\
-          else True
-          
-    @property
-    def nobjects(self):
-        if not self.has_data():
-            return None
-        return len(self.data)
-    
-    # ----------------
-    # - WCS
-    @property
-    def wcs(self):
-        return self._side_properties['wcs']
-
-    def has_wcs(self):
-        return False if self.wcs is None else True
-    
-    @property
-    def radec(self):
-        if not self.has_wcs():
-            raise AttributeError("no 'wcs' solution avialable. Cannot have radec")
-        return np.asarray(self.wcs.pix2world(self.data["x"],
-                                           self.data["y"])).T
-
-    @property
-    def xy(self):
-        return np.asarray([self.get("x"),self.get("y")])
-
-    @property
-    def sky_radec(self):
-        """This is an advanced radec methods tight to astropy SkyCoords"""
-        ra,dec = self.radec
-        return coordinates.SkyCoord(ra=ra*units.degree,dec=dec*units.degree)
-    
-    # ----------------      
-    # -- CATALOGUE
-    @property
-    def catalogue(self):
-        return self._side_properties["catalogue"]
-    
-    def has_catalogue(self):
-        return False if self.catalogue is None\
-          else True
-        
-    @property
-    def catmask(self):
-        """this array tells you if sextractor objects have a catalogue match"""
-        if not self.has_catmatch() and self.has_data():
-            return np.ones(self.nobjects,dtype=bool)
-        return self._derived_properties["catmatch"]["idx"]
-
-    @property
-    def starmask(self):
-        """ This is which are stars amoung *ALL* the detected sources, even those without matcatching"""
-        return self._derived_properties["starmask"]
-    
-
-    @property
-    def isolatedmask(self):
-        return self.catalogue.isolatedmask[self.catmcatch["idx_catalogue"]]
-    
-    @property
-    def catstarmask(self):
-        """return a bool mask of the stars. If it can not do it, not data are masked (full True)"""
-        if not self.has_catmatch() and self.has_data():
-            return np.ones(self.nobjects,dtype=bool)
-        
-        if not self.catalogue.has_starmask():
-            return np.ones(self.nobjects,dtype=bool)
-        
-        return self.catalogue.starmask[ self.catmatch["idx_catalogue"]]
-
-    @property
-    def catisolatedmask(self):
-        """return a bool mask of the stars. If it can not do it, not data are masked (full True)"""
-        if not self.has_catmatch() and self.has_data():
-            return np.ones(self.nobjects,dtype=bool)
-        
-        if not self.catalogue.has_starmask():
-            return np.ones(self.nobjects,dtype=bool)
-        
-        return self.catalogue.isolatedmask[ self.catmatch["idx_catalogue"]]
-    
-    
-    # ----------------------
-    # - Catalogue Matching
-    @property
-    def catmatch(self):
-        """This is the match dictionnary"""
-        if self._derived_properties["catmatch"] is None:
-            self._derived_properties["catmatch"] = {}
-            
-        return self._derived_properties["catmatch"]
-
-    
-    def has_catmatch(self):
-        return False if self.catmatch is None or len(self.catmatch.keys())==0 \
-          else True
-
